requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
chardet==3.0.4
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
Django==3.2.25
django-debug-toolbar==6.0.0
django-jazzmin==.5.0
django-parler==2.3
flatbuffers==25.9.23
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.1
openai==2.5.0
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pip==.2
platformdirs==4.5.0
pooch==1.8.2
protobuf==6.33.0
psycopg2-binary==2.9.3
pydantic==2.12.3
pydantic_core==2.41.4
PyMatting==1.1.14
pyreadline3==3.5.4
python-decouple==3.6
python-dotenv==1.1.1
pytz==2025.2
referencing==0.37.0
rembg==2.0.67
requests==2.32.5
rfc3986==1.5.0
rpds-py==0.27.1
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.10.16
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
webcolors==24.11.1
whitenoise==6.2.0